+++ dart/example/main.dart
@@ -3,6 +3,7 @@
 // found in the LICENSE file.
 
 import 'dart:async';
+import 'dart:io';
 
 import 'package:sentry/sentry.dart';
 
@@ -17,9 +18,12 @@ Future<void> main() async {
   SentryEvent processTagEvent(SentryEvent event, Object hint) =>
       event..tags.addAll({'page-locale': 'en-us'});
 
-  await Sentry.init((options) => options
-    ..dsn = dsn
-    ..addEventProcessor(processTagEvent));
+  await Sentry.init(
+    (options) => options
+      ..dsn = dsn
+      ..addEventProcessor(processTagEvent),
+    runApp,
+  );
 
   Sentry.addBreadcrumb(
     Breadcrumb(
@@ -48,7 +52,9 @@ Future<void> main() async {
       ..setTag('build', '579')
       ..setExtra('company-name', 'Dart Inc');
   });
+}
 
+void runApp() async {
   print('\nReporting a complete event example: ');
 
   // Sends a full Sentry event payload to show the different parts of the UI.
@@ -82,6 +88,8 @@ Future<void> main() async {
   } finally {
     await Sentry.close();
   }
+
+  exit(0);
 }
 
 Future<void> loadConfig() async {