+++ dart/test/default_integrations_test.dart
@@ -1,5 +1,6 @@
 import 'package:mockito/mockito.dart';
-import 'package:sentry/sentry.dart';
+import 'package:sentry/sentry.dart' hide isolateErrorIntegration;
+import 'package:sentry/src/isolate_error_integration.dart';
 import 'package:test/test.dart';
 
 import 'mocks.dart';
@@ -11,49 +12,69 @@ void main() {
     fixture = Fixture();
   });
 
-  test('Isolate error adds integration', () async {
-    isolateErrorIntegration(fixture.hub, fixture.options);
+  test(
+    'Isolate error adds integration',
+    () async {
+      isolateErrorIntegration(fixture.hub, fixture.options);
 
-    expect(true,
-        fixture.options.sdk.integrations.contains('isolateErrorIntegration'));
-  });
-
-  test('Isolate error capture errors', () async {
-    final throwable = StateError('error');
-    final stackTrace = StackTrace.current;
-    final error = [throwable, stackTrace];
-
-    // we could not find a way to trigger an error to the current Isolate
-    // and unit test its error handling, so instead we exposed the method,
-    // that handles and captures it.
-    await handleIsolateError(fixture.hub, fixture.options, error);
-
-    final event = verify(
-      await fixture.hub
-          .captureEvent(captureAny, stackTrace: captureAnyNamed('stackTrace')),
-    ).captured.first as SentryEvent;
-
-    expect(SentryLevel.fatal, event.level);
-
-    final throwableMechanism = event.throwable as ThrowableMechanism;
-    expect('isolateError', throwableMechanism.mechanism.type);
-    expect(true, throwableMechanism.mechanism.handled);
-    expect(throwable, throwableMechanism.throwable);
-  });
-
-  test('Run zoned guarded adds integration', () async {
-    isolateErrorIntegration(fixture.hub, fixture.options);
-
-    void callback() {}
-    final integration = runZonedGuardedIntegration(callback);
-
-    await integration(fixture.hub, fixture.options);
-
-    expect(
+      expect(
         true,
-        fixture.options.sdk.integrations
-            .contains('runZonedGuardedIntegration'));
-  });
+        fixture.options.sdk.integrations.contains('isolateErrorIntegration'),
+      );
+    },
+    onPlatform: {
+      'browser': Skip(),
+    },
+  );
+
+  test(
+    'Isolate error capture errors',
+    () async {
+      final throwable = StateError('error');
+      final stackTrace = StackTrace.current;
+      final error = [throwable, stackTrace];
+
+      // we could not find a way to trigger an error to the current Isolate
+      // and unit test its error handling, so instead we exposed the method,
+      // that handles and captures it.
+      await handleIsolateError(fixture.hub, fixture.options, error);
+
+      final event = verify(
+        await fixture.hub.captureEvent(captureAny,
+            stackTrace: captureAnyNamed('stackTrace')),
+      ).captured.first as SentryEvent;
+
+      expect(SentryLevel.fatal, event.level);
+
+      final throwableMechanism = event.throwable as ThrowableMechanism;
+      expect('isolateError', throwableMechanism.mechanism.type);
+      expect(true, throwableMechanism.mechanism.handled);
+      expect(throwable, throwableMechanism.throwable);
+    },
+    onPlatform: {
+      'browser': Skip(),
+    },
+  );
+
+  test(
+    'Run zoned guarded adds integration',
+    () async {
+      isolateErrorIntegration(fixture.hub, fixture.options);
+
+      void callback() {}
+      final integration = runZonedGuardedIntegration(callback);
+
+      await integration(fixture.hub, fixture.options);
+
+      expect(
+          true,
+          fixture.options.sdk.integrations
+              .contains('runZonedGuardedIntegration'));
+    },
+    onPlatform: {
+      'browser': Skip(),
+    },
+  );
 
   test('Run zoned guarded calls callback', () async {
     isolateErrorIntegration(fixture.hub, fixture.options);
@@ -68,7 +89,7 @@ void main() {
     await integration(fixture.hub, fixture.options);
 
     expect(true, called);
-  });
+  }, onPlatform: {'browser': Skip()});
 
   test('Run zoned guarded calls catches error', () async {
     final throwable = StateError('error');