+++ dart/test/sentry_test.dart
@@ -4,6 +4,8 @@ import 'package:test/test.dart';
 
 import 'mocks.dart';
 
+Function callback = () {};
+
 void main() {
   group('Sentry capture methods', () {
     SentryClient client;
@@ -11,7 +13,7 @@ void main() {
     Exception anException;
 
     setUp(() async {
-      await Sentry.init((options) => options.dsn = fakeDsn);
+      await Sentry.init((options) => options.dsn = fakeDsn, callback);
       anException = Exception('anException');
 
       client = MockSentryClient();
@@ -60,19 +62,28 @@ void main() {
   group('Sentry is enabled or disabled', () {
     test('null DSN', () {
       expect(
-        () async => await Sentry.init((options) => options.dsn = null),
+        () async =>
+            await Sentry.init((options) => options.dsn = null, callback),
+        throwsArgumentError,
+      );
+      expect(Sentry.isEnabled, false);
+    });
+
+    test('null appRunner', () {
+      expect(
+        () async => await Sentry.init((options) => options.dsn = fakeDsn, null),
         throwsArgumentError,
       );
       expect(Sentry.isEnabled, false);
     });
 
     test('empty DSN', () async {
-      await Sentry.init((options) => options.dsn = '');
+      await Sentry.init((options) => options.dsn = '', callback);
       expect(Sentry.isEnabled, false);
     });
 
     test('close disables the SDK', () async {
-      await Sentry.init((options) => options.dsn = fakeDsn);
+      await Sentry.init((options) => options.dsn = fakeDsn, callback);
 
       Sentry.bindClient(MockSentryClient());
 
@@ -93,17 +104,48 @@ void main() {
       var called = false;
       void integration(Hub hub, SentryOptions options) => called = true;
 
-      await Sentry.init((options) {
-        options.dsn = fakeDsn;
-        options.addIntegration(integration);
-      });
+      await Sentry.init(
+        (options) {
+          options.dsn = fakeDsn;
+          options.addIntegration(integration);
+        },
+        callback,
+      );
 
       expect(called, true);
     });
+
+    test('should add default integrations', () async {
+      await Sentry.init(
+        (options) {
+          options.dsn = fakeDsn;
+          expect(
+            options.integrations.contains(isolateErrorIntegration),
+            true,
+          );
+          expect(options.integrations.length, 2);
+        },
+        callback,
+      );
+    }, onPlatform: {'browser': Skip()});
+
+    test('should add only web compatible default integrations', () async {
+      await Sentry.init(
+        (options) {
+          options.dsn = fakeDsn;
+          expect(options.integrations.length, 1);
+        },
+        callback,
+      );
+    }, onPlatform: {'vm': Skip()});
   });
 
-  test("options can't be null", () {
-    expect(() async => await Sentry.init((options) => options = null),
-        throwsArgumentError);
-  });
+  test(
+    "options can't be null",
+    () {
+      expect(
+          () async => await Sentry.init((options) => options = null, callback),
+          throwsArgumentError);
+    },
+  );
 }